+++ src/main/java/com/etl/common/exception/CommonException.java
@@ -1,13 +1,14 @@
 package com.etl.common.exception;
 
 import com.etl.exception.EtlException;
+import com.etl.exception.EtlErrorCategory;
 
 public class CommonException extends EtlException {
     public CommonException(String message) {
-        super(message);
+		super(EtlErrorCategory.RUNTIME, message);
     }
 
     public CommonException(String message, Throwable cause) {
-        super(message, cause);
+		super(EtlErrorCategory.RUNTIME, message, cause);
     }
 }