+++ src/main/java/com/etl/config/exception/ConfigException.java
@@ -1,12 +1,13 @@
 package com.etl.config.exception;
 
 import com.etl.exception.EtlException;
+import com.etl.exception.EtlErrorCategory;
 
 public class ConfigException extends EtlException {
     public ConfigException(String message) {
-        super(message);
+		super(EtlErrorCategory.CONFIG, message);
     }
     public ConfigException(String message, Throwable cause) {
-        super(message, cause);
+		super(EtlErrorCategory.CONFIG, message, cause);
     }
 }