+++ src/main/java/com/etl/processor/exception/ProcessorException.java
@@ -1,13 +1,14 @@
 package com.etl.processor.exception;
 
 import com.etl.exception.EtlException;
+import com.etl.exception.EtlErrorCategory;
 
 public class ProcessorException extends EtlException {
     public ProcessorException(String message) {
-        super(message);
+		super(EtlErrorCategory.FACTORY, message);
     }
 
     public ProcessorException(String message, Throwable cause) {
-        super(message, cause);
+		super(EtlErrorCategory.FACTORY, message, cause);
     }
 }